research/mongodb/filler/requirements.txt
psycopg2-binary==2.9.1
pymongo==4.3.2